Fixed online status checking
Moved the online status check from the forloop to outside of the
forloop. This solved an issue where elements weren't getting properly
organised, as the forloop didn't account for the reduction in the
List<User> size, after the removal of an entry. A seperate forloop now
removes offline users, then passes the main forloop the modified list.
--Note: Incorporate this fix as part of the UserVoiceMuteParser